+++ CHANGELOG.md
@@ -72,6 +72,7 @@
   needlessly indented if preceded by a comment.
 - Line endings other than `\n` are now handled by the formatter, preserving
   blank lines and converting them to `\n`.
+- Fixed a bug where comments would be moved out of an empty list.
 
 ### Build tool